AsyncWorker: make callback optional
`AsyncWorker` assumed that after work is complete,  a JavaScript
 callback would need to execute.

This change removes the restriction of specifying a `Function` callback,
and instead replaces it with an `Env` parameter. Since the purpose of
`receiver` was for the `this` context for the callback, it has also been
removed from the constructors.

Re: https://github.com/nodejs/node-addon-api/issues/231#issuecomment-483356091

PR-URL: https://github.com/nodejs/node-addon-api/pull/489
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>